+++ rl_zoo/train.py
@@ -7,7 +7,6 @@ import uuid
 
 import gym
 import numpy as np
-import seaborn
 import torch as th
 from stable_baselines3.common.utils import set_random_seed
 
@@ -15,8 +14,6 @@ from stable_baselines3.common.utils import set_random_seed
 import rl_zoo.import_envs  # noqa: F401 pytype: disable=import-error
 from rl_zoo.exp_manager import ExperimentManager
 from rl_zoo.utils import ALGOS, StoreDict
-
-seaborn.set()
 
 
 def train():