+++ rl_zoo/wrappers.py
@@ -1,7 +1,6 @@
 import gym
 import numpy as np
 from sb3_contrib.common.wrappers import TimeFeatureWrapper  # noqa: F401 (backward compatibility)
-from scipy.signal import iirfilter, sosfilt, zpk2sos
 
 
 class DoneOnSuccessWrapper(gym.Wrapper):
@@ -53,70 +52,6 @@ class ActionNoiseWrapper(gym.Wrapper):
         noise = np.random.normal(np.zeros_like(action), np.ones_like(action) * self.noise_std)
         noisy_action = action + noise
         return self.env.step(noisy_action)
-
-
-# from https://docs.obspy.org
-def lowpass(data, freq, df, corners=4, zerophase=False):
-    """
-    Butterworth-Lowpass Filter.
-
-    Filter data removing data over certain frequency ``freq`` using ``corners``
-    corners.
-    The filter uses :func:`scipy.signal.iirfilter` (for design)
-    and :func:`scipy.signal.sosfilt` (for applying the filter).
-
-    :type data: numpy.ndarray
-    :param data: Data to filter.
-    :param freq: Filter corner frequency.
-    :param df: Sampling rate in Hz.
-    :param corners: Filter corners / order.
-    :param zerophase: If True, apply filter once forwards and once backwards.
-        This results in twice the number of corners but zero phase shift in
-        the resulting filtered trace.
-    :return: Filtered data.
-    """
-    fe = 0.5 * df
-    f = freq / fe
-    # raise for some bad scenarios
-    if f > 1:
-        f = 1.0
-        msg = "Selected corner frequency is above Nyquist. " + "Setting Nyquist as high corner."
-        print(msg)
-    z, p, k = iirfilter(corners, f, btype="lowpass", ftype="butter", output="zpk")
-    sos = zpk2sos(z, p, k)
-    if zerophase:
-        firstpass = sosfilt(sos, data)
-        return sosfilt(sos, firstpass[::-1])[::-1]
-    else:
-        return sosfilt(sos, data)
-
-
-class LowPassFilterWrapper(gym.Wrapper):
-    """
-    Butterworth-Lowpass
-
-    :param env:
-    :param freq: Filter corner frequency.
-    :param df: Sampling rate in Hz.
-    """
-
-    def __init__(self, env: gym.Env, freq: float = 5.0, df: float = 25.0):
-        super().__init__(env)
-        self.freq = freq
-        self.df = df
-        self.signal = []
-
-    def reset(self):
-        self.signal = []
-        return self.env.reset()
-
-    def step(self, action):
-        self.signal.append(action)
-        filtered = np.zeros_like(action)
-        for i in range(self.action_space.shape[0]):
-            smoothed_action = lowpass(np.array(self.signal)[:, i], freq=self.freq, df=self.df)
-            filtered[i] = smoothed_action[-1]
-        return self.env.step(filtered)
 
 
 class ActionSmoothingWrapper(gym.Wrapper):